test/requirements.txt
pytest==8.3.4
cocotb==1.9.2
pyserial==3.5
